signup fixture and login to bekraeft-konto
